requirements.txt
numpy==1.22.1
Pillow==9.0.1
streamlit==1.6.0
